backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
google-cloud-bigquery==3.10.0